+++ packages/common/src/util/references.js
@@ -19,6 +19,10 @@ const isStream = value => {
 };
 
 function expandReference(state, value) {
+  if (Buffer.isBuffer(value)) {
+    return value;
+  }
+
   if (Array.isArray(value)) {
     return value.map(v => expandReference(state, v));
   }