+++ packages/mailchimp/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @openfn/language-mailchimp
 
+## 0.7.2
+
+### Patch Changes
+
+- 1131c34: Remove regex pattern for validation and changed minLength to 1
+
+## 0.7.1
+
+### Patch Changes
+
+- 1f856c4: Update configuration schema
+- 48394f5: - fix ast docs warnings
+  - add status code log on request
+
 ## 0.7.0
 
 ### Minor Changes