+++ packages/nexmo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @openfn/language-nexmo
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies [c19efbe]
+  - @openfn/language-common@1.11.1
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [85c35b8]
+  - @openfn/language-common@1.11.0
+
 ## 0.4.1
 
 ### Patch Changes