+++ packages/primero/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @openfn/language-primero
 
+## 2.11.8
+
+### Patch Changes
+
+- Updated dependencies [c19efbe]
+  - @openfn/language-common@1.11.1
+
+## 2.11.7
+
+### Patch Changes
+
+- Updated dependencies [85c35b8]
+  - @openfn/language-common@1.11.0
+
 ## 2.11.6
 
 ### Patch Changes